+++ tests/sql_test_files/data/variables_function_get_and_set_duplicate.sql
@@ -1,0 +1,2 @@
+set x = '2';
+select '{"x":"2"}' as expected, sqlpage.variables() as actual;